Added get_str functions
