+++ .gitignore
@@ -1,40 +1,1 @@
-/cmake-build-debug/.ninja_deps
-/cmake-build-debug/.ninja_log
-/cmake-build-debug/CMakeFiles/3.21.1/CompilerIdC/a.out
-/cmake-build-debug/build.ninja
-/cmake-build-debug/.cmake/api/v1/query/cache-v2
-/cmake-build-debug/.cmake/api/v1/reply/cache-v2-d09173c209d18e64c075.json
-/cmake-build-debug/Calc
-/cmake-build-debug/CMakeFiles/clion-environment.txt
-/cmake-build-debug/CMakeFiles/clion-log.txt
-/cmake-build-debug/CMakeFiles/cmake.check_cache
-/cmake-build-debug/cmake_install.cmake
-/cmake-build-debug/CMakeCache.txt
-/cmake-build-debug/CMakeFiles/3.21.1/CMakeCCompiler.cmake
-/cmake-build-debug/CMakeFiles/3.21.1/CompilerIdC/CMakeCCompilerId.c
-/cmake-build-debug/CMakeFiles/3.21.1/CMakeDetermineCompilerABI_C.bin
-/cmake-build-debug/.cmake/api/v1/query/cmakeFiles-v1
-/cmake-build-debug/.cmake/api/v1/reply/cmakeFiles-v1-4f211fc6a4e2bdd606c1.json
-/cmake-build-debug/CMakeFiles/CMakeOutput.log
-/cmake-build-debug/CMakeFiles/3.21.1/CMakeSystem.cmake
-/cmake-build-debug/.cmake/api/v1/query/codemodel-v2
-/cmake-build-debug/.cmake/api/v1/reply/codemodel-v2-b5e9dc4643db86b30749.json
-/cmake-build-debug/.cmake/api/v1/reply/directory-.-Debug-f5ebdc15457944623624.json
-/cmake-build-debug/.cmake/api/v1/reply/index-2022-03-03T14-14-52-0867.json
-/cmake-build-debug/Testing/Temporary/LastTest.log
-/cmake-build-debug/CMakeFiles/Calc.dir/src/main.c.o
-/cmake-build-debug/CMakeFiles/rules.ninja
-/cmake-build-debug/.cmake/api/v1/reply/target-Calc-Debug-6b225a7a3ada23a4ed5c.json
-/cmake-build-debug/CMakeFiles/TargetDirectories.txt
-/cmake-build-debug/.cmake/api/v1/query/toolchains-v1
-/cmake-build-debug/.cmake/api/v1/reply/toolchains-v1-4c4d7253bfc746b21cff.json
-/cmake-build-debug/.cmake/api/v1/reply/cmakeFiles-v1-39534f7622c3037e0f52.json
-/cmake-build-debug/.cmake/api/v1/reply/codemodel-v2-0a1a38c0402ab0fb2ab3.json
-/cmake-build-debug/.cmake/api/v1/reply/index-2022-03-03T17-26-27-0557.json
-/cmake-build-debug/CMakeFiles/Calc.dir/src/operations.c.o
-/cmake-build-debug/CMakeFiles/Calc.dir/src/stack.c.o
-/cmake-build-debug/.cmake/api/v1/reply/target-Calc-Debug-ed59bb21c82dfd6f45bf.json
-/cmake-build-debug/.cmake/api/v1/reply/codemodel-v2-c542cb03aca139667729.json
-/cmake-build-debug/.cmake/api/v1/reply/index-2022-03-03T17-44-53-0065.json
-/src/build/main
-/cmake-build-debug/.cmake/api/v1/reply/target-Calc-Debug-a93623c145d52a61cc5b.json
+/cmake-build-debug/
