+++ src/get_str.h
@@ -1,0 +1,4 @@
+#pragma once
+
+char *get_string_from_input();
+char *get_new_str();
